rename create temp reference function
