+++ app/routes/signin-oidc.js
@@ -10,7 +10,7 @@ const { NoApplicationFoundError, InvalidPermissionsError, ClaimHasAlreadyBeenMad
 const { raiseIneligibilityEvent } = require('../event')
 const { changeContactHistory } = require('../api-requests/contact-history-api')
 const appInsights = require('applicationinsights')
-const createClaimReference = require('../lib/create-claim-reference')
+const createClaimReference = require('../lib/create-temp-claim-reference')
 
 module.exports = [{
   method: 'GET',
